黑名单重复记录：b_shuhe_black_phone_record

+++ marketing-cs/src/main/java/com/br/marketing/mapper/PhoneSaleExtendInfoMapperBase.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PhoneSaleExtendInfoMapperBase {
-    long countByExample(PhoneSaleExtendInfoExample example);
+    int countByExample(PhoneSaleExtendInfoExample example);
 
     int deleteByExample(PhoneSaleExtendInfoExample example);
